fix instrumentation in tests

+++ src/main/java/net/devtech/grossfabrichacks/instrumentation/InstrumentationApi.java
@@ -184,12 +184,14 @@ public class InstrumentationApi {
     }
 
     static {
-        final File agent = GrossFabricHacks.getAgent();
-        final String name = ManagementFactory.getRuntimeMXBean().getName();
+        if (InstrumentationAgent.instrumentation == null) {
+            final File agent = GrossFabricHacks.Common.getAgent();
+            final String name = ManagementFactory.getRuntimeMXBean().getName();
 
-        ByteBuddyAgent.attach(agent, name.substring(0, name.indexOf('@')));
+            ByteBuddyAgent.attach(agent, name.substring(0, name.indexOf('@')));
 
-        agent.delete();
+            agent.delete();
+        }
 
         instrumentation = InstrumentationAgent.instrumentation;
     }
